refactor(auth): split modalSingIn
