+++ src/containers/ModalSignIn/DefinePassword/index.ts
@@ -1,0 +1,2 @@
+export * from './DefinePasswordField'
+export * from './useDefinePasswordStep'
